Wrote up the basic bits for activation and damage.
Units are instantiated with a list of damageResponders, each of which is a damageable. Those all take damage for the unit until they can't anymore, and then the unit's health takes damage.
Units are also instantiated with a list of activationResponders, each of which is an activatable. When you activate the unit, it activates all its activatables. THIS BIT IS CURRENTLY UNTESTED.

Implemented DamageBar, which tracks a health-style resource which is initialized with a max and bottoms out at zero.
Implemented Health using DamageBar. Only additional functionality is that it can tell if you're dead or not.
Implemented Shields using DamageBar. They recharge after you get activated enough times without taking damage.

Damageable extension includes two convenience damage() funcs that initialize up a Damage object for you if given an amount, and, optionally, a type.

To access a Unit's health, shields, maxHealth, whether they're dead or not, use `attributes[{attributeName}]`